*PlayerStat.cs, *PlayerBattle.cs, *UIBar.cs
플레이어 체력 float에서 int형으로 수정
마나 수치 반영

+++ RedCloak/Assets/Scripts/Player/PlayerBattle.cs
@@ -45,7 +45,7 @@ public class PlayerBattle : MonoBehaviour
         }
     }
 
-    public bool ChangeHealth(float change)
+    public bool ChangeHealth(int change)
     {
         if (timeSinceLastChange < healthChangeDelay) // if not attacked
         {
